moved getTypographyTokens
